Update accordion component js and css
