+++ src/scripts/main.js
@@ -6,3 +6,4 @@ import './progressbar.js';
 import './snackbar.js';
 import './marcRecordUi.js';
 import './panelTabs.js';
+import './accordion.js'
